wedding-backend/requirements.txt
asgiref==3.5.0
astroid==2.3.3
boto3==1.12.31
botocore==1.15.31
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.3
Django==3.2
django-cors-headers==3.11.0
django-health-check==3.16.5
django-storages==1.12.3
djangorestframework==3.13.1
docutils==0.15.2
drfpasswordless==1.5.8
gunicorn==20.1.0
idna==3.4
isort==4.3.21
jmespath==0.9.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==9.0.1
psycopg2==2.9.3
pylint==2.4.4
pylint-django==2.0.14
pylint-plugin-utils==0.6
python-dateutil==2.8.1
pytz==2022.1
requests==2.28.1
s3transfer==0.3.3
six==1.14.0
sqlparse==0.4.2
urllib3==1.25.8
wrapt==1.11.2
